libmints now cleanly compiles in Debug configuration

+++ src/lib/libmints/wavefunction.cc
@@ -421,12 +421,12 @@ SharedMatrix Wavefunction::C_subset_helper(SharedMatrix C, const Dimension& nocc
     std::vector<std::vector<int> > positions = subset_occupation(noccpi, subset);
 
     Dimension nmopi(nirrep_);
-    for (int h = 0; h < positions.size(); h++) {
+    for (int h = 0; h < (int)positions.size(); h++) {
         nmopi[h] = positions[h].size();
     }
     SharedMatrix C2(new Matrix("C " + basis + " " + subset, nsopi_, nmopi));
-    for (int h = 0; h < positions.size(); h++) {
-        for (int i = 0; i < positions[h].size(); i++) {
+    for (int h = 0; h < (int)positions.size(); h++) {
+        for (int i = 0; i < (int)positions[h].size(); i++) {
             C_DCOPY(nsopi_[h], &C->pointer(h)[0][positions[h][i]], nmopi_[h], &C2->pointer(h)[0][i], nmopi[h]);
         }
     }
@@ -438,14 +438,14 @@ SharedMatrix Wavefunction::C_subset_helper(SharedMatrix C, const Dimension& nocc
 
         std::vector<boost::tuple<double, int, int> > order;
         for (int h = 0; h < nirrep_; h++) {
-            for (int i = 0; i < positions[h].size(); i++) {
+            for (int i = 0; i < (int)positions[h].size(); i++) {
                 order.push_back(boost::tuple<double,int,int>(epsilon->get(h,positions[h][i]),i,h));
             }
         }
 
         std::sort(order.begin(), order.end(), std::less<boost::tuple<double,int,int> >());
 
-        for (int index = 0; index < order.size(); index++) {
+        for (int index = 0; index < (int)order.size(); index++) {
             int i = boost::get<1>(order[index]);
             int h = boost::get<2>(order[index]);
 
@@ -470,7 +470,7 @@ SharedVector Wavefunction::epsilon_subset_helper(SharedVector epsilon, const Dim
     std::vector<std::vector<int> > positions = subset_occupation(noccpi, subset);
 
     Dimension nmopi(nirrep_);
-    for (int h = 0; h < positions.size(); h++) {
+    for (int h = 0; h < (int)positions.size(); h++) {
         nmopi[h] = positions[h].size();
     }
 
@@ -482,22 +482,22 @@ SharedVector Wavefunction::epsilon_subset_helper(SharedVector epsilon, const Dim
 
         std::vector<boost::tuple<double, int, int> > order;
         for (int h = 0; h < nirrep_; h++) {
-            for (int i = 0; i < positions[h].size(); i++) {
+            for (int i = 0; i < (int)positions[h].size(); i++) {
                 order.push_back(boost::tuple<double,int,int>(epsilon->get(h,positions[h][i]),i,h));
             }
         }
 
         std::sort(order.begin(), order.end(), std::less<boost::tuple<double,int,int> >());
 
-        for (int index = 0; index < order.size(); index++) {
+        for (int index = 0; index < (int)order.size(); index++) {
             C2->set(0,index,boost::get<0>(order[index]));
         }
 
     } else if (basis == "SO" || basis == "MO") {
 
         C2 = SharedVector(new Vector("Epsilon " + basis + " " + subset, nmopi));
-        for (int h = 0; h < positions.size(); h++) {
-            for (int i = 0; i < positions[h].size(); i++) {
+        for (int h = 0; h < (int)positions.size(); h++) {
+            for (int i = 0; i < (int)positions[h].size(); i++) {
                 C2->set(h,i,epsilon->get(h,positions[h][i]));
             }
         }
@@ -817,11 +817,11 @@ void Wavefunction::CIMSet(bool value,int nactive_occupied)
     CIM_orbital_factors_.reset();
     CIM_orbital_energies_.reset();
 
-    QLMO_to_LMO_ = 
+    QLMO_to_LMO_ =
         SharedMatrix (new Matrix("CIM Rii",nactive_occupied,nactive_occupied));
-    CIM_orbital_factors_ = 
+    CIM_orbital_factors_ =
         SharedVector (new Vector("CIM Orbital Factors",nactive_occupied));
-    CIM_orbital_energies_ = 
+    CIM_orbital_energies_ =
         SharedVector (new Vector("CIM Orbital Energies",nmo_));
 }
 
@@ -830,7 +830,7 @@ bool Wavefunction::isCIM()
     return isCIM_;
 }
 
-boost::shared_ptr<Vector> Wavefunction::get_atomic_point_charges() const { 
+boost::shared_ptr<Vector> Wavefunction::get_atomic_point_charges() const {
     boost::shared_ptr<double[]> q = atomic_point_charges();
 
     int n = molecule_->natom();
@@ -840,7 +840,7 @@ boost::shared_ptr<Vector> Wavefunction::get_atomic_point_charges() const {
     return q_vector;
 }
 
-void Wavefunction::load_values_from_chkpt() 
+void Wavefunction::load_values_from_chkpt()
 {
 /*
  Below is a list of all quantities that could be stored in the chkpt file as of August 2014.
@@ -859,7 +859,7 @@ void Wavefunction::load_values_from_chkpt()
 //    nalphapi_ = Dimension(nirrep_, "Alpha electrons per irrep");
 //    nbetapi_  = Dimension(nirrep_, "Beta electrons per irrep");
 
-    int *v; 
+    int *v;
     if (chkpt_->exist_add_prefix("SO's per irrep")) {
       v = chkpt_->rd_sopi();
       for (int i=0; i<nirrep_; ++i) nsopi_[i] = v[i];
